switch to floats

+++ network/layer.hpp
@@ -22,13 +22,13 @@ class Layer {
     public:
     //==============================================
     Layer(int N, int M, std::string activation_function)
-    : _W(Eigen::MatrixXd::Random(M, N)) // randomly initialize weight matrix
-    , _b(Eigen::VectorXd::Zero(M)) // bias set to 0 vector
-    , _x(Eigen::VectorXd::Zero(N)) // input set to 0
-    , _z(Eigen::VectorXd::Zero(M)) // output before activation. store for gradient computations
-    , _y(Eigen::VectorXd::Zero(M)) // output after activation
-    , _dLdW(Eigen::MatrixXd::Zero(M, N)) // gradient w.r.t. weight matrix
-    , _dLdb(Eigen::VectorXd::Zero(M)) // gradient w.r.t. bias vector
+    : _W(Eigen::MatrixXf::Random(M, N)) // randomly initialize weight matrix
+    , _b(Eigen::VectorXf::Zero(M)) // bias set to 0 vector
+    , _x(Eigen::VectorXf::Zero(N)) // input set to 0
+    , _z(Eigen::VectorXf::Zero(M)) // output before activation. store for gradient computations
+    , _y(Eigen::VectorXf::Zero(M)) // output after activation
+    , _dLdW(Eigen::MatrixXf::Zero(M, N)) // gradient w.r.t. weight matrix
+    , _dLdb(Eigen::VectorXf::Zero(M)) // gradient w.r.t. bias vector
     , _activation_name(activation_function)
     {
         if (activation_function == "ReLU") {
@@ -63,20 +63,20 @@ class Layer {
     //==============================================
     // getters
     //==============================================
-    Eigen::VectorXd get_x() { return _x; }
-    Eigen::VectorXd get_z() { return _z; }
-    Eigen::VectorXd get_y() { return _y; }
-    Eigen::MatrixXd get_W() { return _W; }
-    Eigen::VectorXd get_b() { return _b; }
-    Eigen::VectorXd get_dldb() { return _dLdb; }
-    Eigen::MatrixXd get_dldW() { return _dLdW; }
+    Eigen::VectorXf get_x() { return _x; }
+    Eigen::VectorXf get_z() { return _z; }
+    Eigen::VectorXf get_y() { return _y; }
+    Eigen::MatrixXf get_W() { return _W; }
+    Eigen::VectorXf get_b() { return _b; }
+    Eigen::VectorXf get_dldb() { return _dLdb; }
+    Eigen::MatrixXf get_dldW() { return _dLdW; }
 
 
     //==============================================
     // update step
     // updates the weights and biases of the layer
     //==============================================
-    void update(double lr) {
+    void update(float lr) {
         _W.noalias() -= lr * _dLdW;
         _b.noalias() -= lr * _dLdb;
     }
@@ -87,7 +87,7 @@ class Layer {
     // returns an output vector after applying the activation function layer-wise
     // y = a(Wx+b)
     //==============================================
-    Eigen::VectorXd forward(const Eigen::Ref<const Eigen::VectorXd>& x) {
+    Eigen::VectorXf forward(const Eigen::Ref<const Eigen::VectorXf>& x) {
         assert(x.size() == _W.cols() && "Input vector size must match weight matrix columns");
         _x = x; // store input
         _z = _W * _x + _b; // store linear output
@@ -102,10 +102,10 @@ class Layer {
     // weight matrix, and bias vector via chain rule. final output is the gradient of the layer output
     // w.r.t the input vector (dL/dx) 
     //==============================================
-    Eigen::VectorXd backward(const Eigen::Ref<const Eigen::VectorXd>& upstream_gradient) {
+    Eigen::VectorXf backward(const Eigen::Ref<const Eigen::VectorXf>& upstream_gradient) {
         assert(upstream_gradient.size() == _y.size() && "Upstream gradient size must match output size");
 
-        const Eigen::VectorXd* derivative_input;
+        const Eigen::VectorXf* derivative_input;
         switch(_derivative_input) {
             case(Z): 
                 derivative_input = &_z;
@@ -117,11 +117,11 @@ class Layer {
                 throw std::logic_error("Unknown derivative input type. This should never happen");
         }
         
-        Eigen::VectorXd dadx = _activation_derivative_function(*derivative_input);
-        Eigen::VectorXd dLdz = upstream_gradient.cwiseProduct(dadx);
-        Eigen::VectorXd dLdx = _W.transpose() * dLdz;
-        Eigen::MatrixXd dLdW = dLdz * _x.transpose();
-        Eigen::VectorXd dLdb = dLdz;
+        Eigen::VectorXf dadx = _activation_derivative_function(*derivative_input);
+        Eigen::VectorXf dLdz = upstream_gradient.cwiseProduct(dadx);
+        Eigen::VectorXf dLdx = _W.transpose() * dLdz;
+        Eigen::MatrixXf dLdW = dLdz * _x.transpose();
+        Eigen::VectorXf dLdb = dLdz;
         _dLdW += dLdW;
         _dLdb += dLdb;
         return dLdx;
@@ -137,21 +137,21 @@ class Layer {
 
     private:
 
-    using ActivationFunction = std::function<Eigen::VectorXd(const Eigen::Ref<const Eigen::VectorXd>&)>;
-    using ActivationDerivativeFunction = std::function<Eigen::VectorXd(const Eigen::Ref<const Eigen::VectorXd>&)>;
+    using ActivationFunction = std::function<Eigen::VectorXf(const Eigen::Ref<const Eigen::VectorXf>&)>;
+    using ActivationDerivativeFunction = std::function<Eigen::VectorXf(const Eigen::Ref<const Eigen::VectorXf>&)>;
     enum DerivativeInput { Z, Y };
 
     ActivationFunction _activation_function;
     ActivationDerivativeFunction _activation_derivative_function;
     DerivativeInput _derivative_input;
 
-    Eigen::MatrixXd _W; // weight matrix
-    Eigen::VectorXd _b; // bias vector
-    Eigen::VectorXd _x; // input vector
-    Eigen::VectorXd _z; // linear output: Wx+b
-    Eigen::VectorXd _y; // activated output: a(Wx+b)
-    Eigen::MatrixXd _dLdW; // gradient w.r.t. weights: dL/dW
-    Eigen::VectorXd _dLdb; // gradient w.r.t. bias: dL/db 
+    Eigen::MatrixXf _W; // weight matrix
+    Eigen::VectorXf _b; // bias vector
+    Eigen::VectorXf _x; // input vector
+    Eigen::VectorXf _z; // linear output: Wx+b
+    Eigen::VectorXf _y; // activated output: a(Wx+b)
+    Eigen::MatrixXf _dLdW; // gradient w.r.t. weights: dL/dW
+    Eigen::VectorXf _dLdb; // gradient w.r.t. bias: dL/db 
     std::string _activation_name;
 };
